Create Stones on the Table.cpp
QUESTION LINK: https://codeforces.com/problemset/problem/266/A

INPUT FORMAT:
The first line contains integer n (1 ≤ n ≤ 50) — the number of stones on the table.

The next line contains string s, which represents the colors of the stones. We'll consider the stones in the row numbered from 1 to n from left to right. Then the i-th character s equals "R", if the i-th stone is red, "G", if it's green and "B", if it's blue.

OUTPUT FORMAT:
Print a single integer — the answer to the problem

SAMPLE INPUT 1:
3
RRG
SAMPLE OUTPUT 1:
1

SAMPLE INPUT 2:
5
RRRRR
SAMPLE OUTPUT 2:
4

SAMPLE INPUT 3:
4
BRBG
SAMPLE OUTPUT 3:
0